+++ test/realworld.spec.ts
@@ -5,11 +5,13 @@ import googleMapsDistanceMatrix from "./dataset/google-maps-distance-matrix.json
 import { lruCacheStat } from "../src/encoder/lru-cache.js"
 import { bufferAllocatorStat } from "../src/encoder/byte-array.js"
 import { prefixTrieStat } from "../src/decoder/prefix-trie.js"
+import { stringBufferStat } from "../src/encoder/string-buffer.js"
 
 afterAll(() => {
   console.log(lruCacheStat())
   console.log(bufferAllocatorStat())
   console.log(prefixTrieStat())
+  console.log(stringBufferStat())
 })
 
 describe("Realworld", () => {